Cyber_Sec/Scan_NMAP/requirement.txt
python-nmap
ftputil
requests
pysmb
fabric
subprocess32
ipstack
pyowm
wikipedia-api
pymongo